refactor(dashboard): extract <Eyebrow> component, split email-identities
Code review follow-up — two CODE_QUALITY.md violations from the
editorial polish PR.

1. Extract <Eyebrow>. The all-caps + wide-tracking label recipe was
   duplicated 21 times across 9 files with slightly different
   variations (size, tone, weight, mono). Now a single component with
   three size variants (micro / sm / md), four tones (muted / brand /
   subtle / bright), and `mono` + `weight` props. Call sites drop
   from 7-token className strings to one element.

   Swapped in: analytics page.tsx (hero + supporting labels + section
   h2s), tasks/[id] page.tsx (4 section headers), system-status.tsx
   (dt rows), email-identities.tsx (transport chip), shell-empty-state
   (language label), form-renderer text + media + complex.

   Leaves two call sites: table `<tr>` headers in audit + page. Those
   are structural (class on the row, cascading to children) and don't
   map to an Eyebrow cleanly.

2. Split email-identities.tsx. 329 lines → 152 + 214. The creation
   form was a second concern inside the list component — now it's in
   email-identity-form.tsx with its own TRANSPORTS + input style + per-
   transport hint map extracted as module constants. The IdentityRow
   also pulled out into its own component for clarity.

Both are mechanical refactors; typecheck clean, screenshots unchanged.
